move consol-ui methods to UI folder
