+++ UI/Session.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Airport_Ticket_Booking
+namespace Airport_Ticket_Booking.UI
 {
     public static class Session
     {
